# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ChatAI API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/categories/{name}/characters": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}